--- src/main/java/com/canvi/hama/ai/request/AiRequest.java
@@ -1,16 +1,0 @@
-package com.canvi.hama.ai.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AiRequest{
-    @NotBlank
-    private String userName;
-
-    @NotBlank
-    private String prompt;
-}
