--- src/main/java/com/canvi/hama/ai/response/GptResponse.java
@@ -1,21 +1,0 @@
-package com.canvi.hama.ai.response;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class GptResponse {
-    private Long statusCode;
-    private Data data;
-
-    @Getter
-    @Setter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class Data {
-        private String gptResult;
-
-    }
-}
