--- src/main/java/com/canvi/hama/diary/entity/Diary.java
@@ -1,36 +1,0 @@
-package com.canvi.hama.diary.entity;
-
-import com.canvi.hama.common.entity.BaseEntity;
-import com.canvi.hama.domain.user.domain.User;
-import lombok.*;
-
-import jakarta.persistence.*;
-
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "diary")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Diary extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id", referencedColumnName = "id", nullable = false)
-    private User user;
-
-    @Column(name = "title", nullable = false)
-    private String title;
-
-    @Lob
-    @Column(name = "content", nullable = false, columnDefinition = "LONGTEXT")
-    private String content;
-
-    @Column(name = "diary_date", nullable = false)
-    private LocalDate diaryDate;
-}
