--- src/main/java/com/canvi/hama/diary/repository/DiaryRepository.java
@@ -1,15 +1,0 @@
-package com.canvi.hama.diary.repository;
-
-import com.canvi.hama.diary.entity.Diary;
-import org.springframework.data.jpa.repository.EntityGraph;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface DiaryRepository extends JpaRepository<Diary, Long> {
-
-    List<Diary> findByUserId(Long userId);
-
-}
