--- src/main/java/com/canvi/hama/diary/request/CommentSaveRequest.java
@@ -1,16 +1,0 @@
-package com.canvi.hama.diary.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class CommentSaveRequest {
-    private Long diaryId;
-    private Long userId;
-    private String comment;
-}
