+++ src/main/java/com/canvi/hama/domain/ai/exception/AiException.java
@@ -1,0 +1,16 @@
+package com.canvi.hama.domain.ai.exception;
+
+import com.canvi.hama.domain.ai.response.AiResponseStatus;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class AiException extends RuntimeException {
+    private AiResponseStatus status;
+
+    public AiException(AiResponseStatus status) {
+        super(status.getMessage());
+        this.status = status;
+    }
+}
