+++ src/main/java/com/canvi/hama/domain/ai/response/DalleResponse.java
@@ -1,0 +1,23 @@
+package com.canvi.hama.domain.ai.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class DalleResponse {
+    private Long statusCode;
+    private Integer created;
+
+    private Data data;
+
+    @Getter
+    @Setter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class Data {
+        private String revised_prompt;
+        private String url;
+    }
+}
