+++ src/main/java/com/canvi/hama/domain/ai/response/GptResponse.java
@@ -1,0 +1,21 @@
+package com.canvi.hama.domain.ai.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class GptResponse {
+    private Long statusCode;
+    private Data data;
+
+    @Getter
+    @Setter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class Data {
+        private String gptResult;
+
+    }
+}
