+++ src/main/java/com/canvi/hama/domain/diary/entity/Comment.java
@@ -1,0 +1,30 @@
+package com.canvi.hama.domain.diary.entity;
+
+import com.canvi.hama.common.entity.BaseEntity;
+import com.canvi.hama.domain.user.domain.User;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Table(name = "comment")
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Comment extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "diary_id", referencedColumnName = "id", nullable = false)
+    private Diary diaryId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", referencedColumnName = "id", nullable = false)
+    private User userId;
+
+    @Column(name = "comment", nullable = false, columnDefinition = "TEXT")
+    private String comment;
+}
