+++ src/main/java/com/canvi/hama/domain/diary/repository/CommentRepository.java
@@ -1,0 +1,14 @@
+package com.canvi.hama.domain.diary.repository;
+
+import com.canvi.hama.domain.diary.entity.Comment;
+import com.canvi.hama.domain.diary.entity.Diary;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    Optional<Comment> findByDiaryId(Diary diaryId);
+
+}
