+++ src/main/java/com/canvi/hama/domain/diary/repository/DiaryRepository.java
@@ -1,0 +1,14 @@
+package com.canvi.hama.domain.diary.repository;
+
+import com.canvi.hama.domain.diary.entity.Diary;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
+
+    List<Diary> findByUserId(Long userId);
+
+}
