+++ src/test/java/com/canvi/hama/diary/controller/ImageTest.java
@@ -1,11 +1,11 @@
 package com.canvi.hama.diary.controller;
 
 import com.canvi.hama.common.security.JwtTokenProvider;
-import com.canvi.hama.diary.entity.Diary;
-import com.canvi.hama.diary.exception.DiaryException;
-import com.canvi.hama.diary.request.DiaryRequest;
-import com.canvi.hama.diary.request.ImageSaveRequest;
-import com.canvi.hama.diary.response.DiaryResponseStatus;
+import com.canvi.hama.domain.diary.entity.Diary;
+import com.canvi.hama.domain.diary.exception.DiaryException;
+import com.canvi.hama.domain.diary.request.DiaryRequest;
+import com.canvi.hama.domain.diary.request.ImageSaveRequest;
+import com.canvi.hama.domain.diary.response.DiaryResponseStatus;
 import com.canvi.hama.domain.auth.dto.LoginRequest;
 import com.canvi.hama.domain.auth.dto.SignupRequest;
 import com.canvi.hama.domain.user.domain.User;
